Create Ocr Editor parts
